Add missing int cast

+++ dags/daily_builder_stats_to_slack.py
@@ -21,7 +21,7 @@ with DAG(
         sql=f"""with past_24h_submissions as (SELECT builder_pubkey, count(builder_pubkey) as blocks FROM mainnet_payload_delivered WHERE inserted_at BETWEEN NOW() - INTERVAL '1 DAYS' AND NOW() GROUP BY builder_pubkey),
 past_24h_top_pubkeys as (select a.builder_pubkey, a.blocks, b.description FROM past_24h_submissions a LEFT JOIN mainnet_blockbuilder b ON a.builder_pubkey = b.builder_pubkey where a.blocks > 10 ORDER BY blocks DESC LIMIT 20),
 total_blocks as (select count(*) as cnt from mainnet_payload_delivered WHERE inserted_at BETWEEN NOW() - INTERVAL '1 DAYS' AND NOW())
-SELECT builder_pubkey, blocks, description FROM past_24h_top_pubkeys
+SELECT builder_pubkey, blocks::int, description FROM past_24h_top_pubkeys
 union all
 select 'others' as builder_pubkey, (select cnt from total_blocks) - (select sum(blocks) from past_24h_top_pubkeys) as blocks, 'others' as description""",
         slack_conn_id=SLACK_CONN_ID,
